src/functions/requirements.txt
alpaca-trade-api==1.2.0
azure-common==1.1.27
azure-core==1.13.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-nspkg==3.0.2
azure-storage-queue==12.1.6
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
google-api-core==1.26.3
google-api-python-client==2.3.0
google-auth==1.30.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
isodate==0.6.0
msgpack==1.0.2
msrest==0.6.21
numpy==1.20.2
oauthlib==3.1.0
packaging==20.9
pandas==1.2.4
protobuf==3.16.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
uritemplate==3.0.1
urllib3==1.26.4
websocket-client==0.59.0
websockets==8.1
